+++ src/compiler/sema_passes.c
@@ -18,6 +18,7 @@ void parent_path(StringSlice *slice)
 	slice->len = 0;
 }
 
+
 void sema_analyse_pass_module_hierarchy(Module *module)
 {
 	const char *name = module->name->module;
@@ -45,6 +46,30 @@ void sema_analyse_pass_module_hierarchy(Module *module)
 	Module *parent_module = compiler_find_or_create_module(path);
 	module->parent_module = parent_module;
 	sema_analyze_stage(parent_module, ANALYSIS_MODULE_HIERARCHY);
+}
+
+
+void sema_analyse_pass_remove_feat_conditionals(Module *module)
+{
+	FOREACH(CompilationUnit *, unit, module->units)
+	{
+		if (unit->feat_attributes && sema_remove_due_to_conditionals(unit->feat_attributes) == BOOL_TRUE)
+		{
+			vec_resize(unit->module_aliases, 0);
+			vec_resize(unit->imports, 0);
+			vec_resize(unit->global_decls, 0);
+			vec_resize(unit->global_cond_decls, 0);
+			FOREACH(Decl *, decl, module->generic_sections)
+			{
+				if (decl->unit == unit)
+				{
+					vec_resize(decl->generic_decl.conditional_decls, 0);
+					vec_resize(decl->generic_decl.decls, 0);
+				}
+			}
+			vec_resize(unit->ct_includes, 0);
+		}
+	}
 }
 
 void sema_analysis_pass_process_imports(Module *module)
@@ -243,7 +268,7 @@ static Decl **sema_load_include(CompilationUnit *unit, Decl *decl)
 	sema_context_init(&context, unit);
 	FOREACH(Attr *, attr, decl->attributes)
 	{
-		if (attr->attr_kind != ATTRIBUTE_IF)
+		if (attr->attr_kind != ATTRIBUTE_IF && attr->attr_kind != ATTRIBUTE_FEAT)
 		{
 			RETURN_PRINT_ERROR_AT(NULL, attr, "Invalid attribute for '$include'.");
 		}
@@ -319,7 +344,7 @@ static Decl **sema_run_exec(CompilationUnit *unit, Decl *decl)
 	sema_context_init(&context, unit);
 	FOREACH(Attr *, attr, decl->attributes)
 	{
-		if (attr->attr_kind != ATTRIBUTE_IF)
+		if (attr->attr_kind != ATTRIBUTE_IF && attr->attr_kind != ATTRIBUTE_FEAT)
 		{
 			RETURN_PRINT_ERROR_AT(NULL, attr, "Invalid attribute for '$exec'.");
 		}
@@ -404,7 +429,7 @@ static Decl **sema_interpret_expand(CompilationUnit *unit, Decl *decl)
 	sema_context_init(&context, unit);
 	FOREACH(Attr *, attr, decl->attributes)
 	{
-		if (attr->attr_kind != ATTRIBUTE_IF)
+		if (attr->attr_kind != ATTRIBUTE_IF && attr->attr_kind != ATTRIBUTE_FEAT)
 		{
 			RETURN_PRINT_ERROR_AT(NULL, attr, "Invalid attribute for '$expand'.");
 		}
@@ -491,7 +516,6 @@ void sema_analysis_pass_process_includes(Module *module)
 	DEBUG_LOG("Pass: Process includes for module '%s'.", module->name->module);
 	FOREACH(CompilationUnit *, unit, module->units)
 	{
-		if (unit->if_attr) continue;
 		// Process all includes
 		sema_process_includes(unit);
 		ASSERT(vec_size(unit->ct_includes) == 0);
@@ -501,24 +525,24 @@ void sema_analysis_pass_process_includes(Module *module)
 }
 
 
-void sema_analysis_pass_process_methods(Module *module, bool process_generic)
+void sema_analysis_pass_process_methods(Module *module)
 {
 	DEBUG_LOG("Pass: Process methods register for module '%s'.", module->name->module);
 	FOREACH(CompilationUnit *, unit, module->units)
 	{
 		SemaContext context;
 		sema_context_init(&context, unit);
-		FOREACH(Decl *, method, process_generic ? unit->generic_methods_to_register : unit->methods_to_register)
+		unsigned size_before = vec_size(unit->methods_to_register);
+		FOREACH(Decl *, method, unit->methods_to_register)
 		{
 			TypeInfo *parent_type_info = decl_find_method_target(method);
-			if (!process_generic && sema_unresolved_type_is_generic(&context, parent_type_info))
+			if (sema_unresolved_type_is_generic(&context, parent_type_info))
 			{
-				vec_add(unit->generic_methods_to_register, method);
+				vec_add(compiler.context.unregistered_method_specializations, method);
 				continue;
 			}
 			if (sema_analyse_method_register(&context, method))
 			{
-				if (method->decl_kind == DECL_ERASED) continue;
 				if (method->decl_kind == DECL_MACRO)
 				{
 					vec_add(unit->macro_methods, method);
@@ -530,21 +554,51 @@ void sema_analysis_pass_process_methods(Module *module, bool process_generic)
 			}
 		}
 		sema_context_destroy(&context);
-		if (process_generic)
-		{
-			vec_resize(unit->generic_methods_to_register, 0);
-		}
-		else
-		{
-			vec_resize(unit->methods_to_register, 0);
-		}
+		ASSERT(size_before == vec_size(unit->methods_to_register));
+		vec_resize(unit->methods_to_register, 0);
 	}
 
 	DEBUG_LOG("Pass finished with %d error(s).", compiler.context.errors_found);
 }
 
+/*
+ *
+ * A specialization is a generic method like:
+ * fn int Foo{int}.bar(&self) { ... }
+ * That is, the method parent is a generic instantiation.
+ */
+void sema_analysis_pass_process_method_specialization(void)
+{
+	DEBUG_LOG("Pass: Process method specializations");
+	SemaContext context;
+	CompilationUnit *unit = NULL;
+	FOREACH(Decl *, method, compiler.context.unregistered_method_specializations)
+	{
+		CompilationUnit *method_unit = method->unit;
+		if (method_unit != unit)
+		{
+			if (unit) sema_context_destroy(&context);
+			unit = method_unit;
+			sema_context_init(&context, unit);
+		}
+		if (sema_analyse_method_register(&context, method))
+		{
+			if (method->decl_kind == DECL_MACRO)
+			{
+				vec_add(unit->macro_methods, method);
+			}
+			else
+			{
+				vec_add(unit->methods, method);
+			}
+		}
+	}
+	if (unit) sema_context_destroy(&context);
+	vec_resize(compiler.context.unregistered_method_specializations, 0);
+	DEBUG_LOG("Pass finished with %d error(s).", compiler.context.errors_found);
+}
 
-void sema_analysis_pass_register_conditional_units(Module *module)
+void sema_analysis_pass_register_conditional_units_and_decls(Module *module)
 {
 	DEBUG_LOG("Pass: Register conditional units for %s", module->name->module);
 	FOREACH(CompilationUnit *, unit, module->units)
@@ -558,6 +612,7 @@ void sema_analysis_pass_register_conditional_units(Module *module)
 		SemaContext context;
 		sema_context_init(&context, unit);
 		if (!if_attr) goto CHECK_LINK;
+		SEMA_DEPRECATED(if_attr, "Top declaration '@if' is deprecated, please use '@feat' instead.");
 		if (vec_size(if_attr->exprs) != 1)
 		{
 			PRINT_ERROR_AT(if_attr, "Expected one parameter.");
@@ -637,24 +692,19 @@ FAIL_CONTEXT:
 		sema_context_destroy(&context);
 		break;
 	}
-	DEBUG_LOG("Pass finished with %d error(s).", compiler.context.errors_found);
-}
-
-void sema_analysis_pass_register_conditional_declarations(Module *module)
-{
 	DEBUG_LOG("Pass: Register conditional declarations for module '%s'.", module->name->module);
 	FOREACH(CompilationUnit *, unit, module->units)
 	{
 		unit->module = module;
 		DEBUG_LOG("Processing %s.", unit->file->name);
-RETRY:;
+		RETRY:;
 		Decl **decls = unit->global_cond_decls;
 		FOREACH(Decl *, decl, decls)
 		{
 			unit_register_optional_global_decl(unit, decl);
 		}
 		vec_resize(decls, 0);
-RETRY_INCLUDES:
+		RETRY_INCLUDES:
 		decls = unit->ct_includes;
 		unit->ct_includes = NULL;
 		register_includes(unit, decls);
@@ -797,12 +847,10 @@ void sema_analysis_pass_decls(Module *module)
 	{
 		SemaContext context;
 		sema_context_init(&context, unit);
-		context.active_scope = (DynamicScope)
-				{
+		context.active_scope = (DynamicScope) {
 					.depth = 0,
 					.label_start = 0,
-					.current_local = 0,
-				};
+					.current_local = 0, };
 		sema_analyse_decls(&context, unit->attributes);
 		sema_analyse_decls(&context, unit->enums);
 		FOREACH(Decl *, decl, unit->types)